refactor: move metrics feature files into src/modules/metrics
